Refactor the routes in hiw own file
